Load links upon start
This commit adds the function of loading commits upon start of the site.
If a user adds a bunch of links to their page they are now saved into the links.json file in the backend.
Upon start, this will now grab those links and load them into the frontend.